Moved distance normalmaps code to its own file
